Localize confirmation display-name helpers
